+++ BetterScp939/BetterScp939.csproj
@@ -33,9 +33,9 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="BetterScp939.cs" />
-    <Compile Include="Configs.cs" />
+    <Compile Include="Config.cs" />
     <Compile Include="Events\PlayerHandler.cs" />
-    <Compile Include="Components\CustomScp939.cs" />
+    <Compile Include="Components\BetterScp939Controller.cs.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
